refactor: split unsupervised pretraining into multiple modules
