+++ personalized_active_learning/algorithms/__init__.py
@@ -4,4 +4,4 @@ Most of experimental features goes here.
 
 """
 
-from .unsupervised_pretraining import IUnsupervisedPretrainer, KmeansPretrainer
+from .unsupervised_pretraining import KmeansPretrainer, IUnsupervisedPretrainer
